requirements.txt.txt
selenium
schedule
python-telegram-bot
pyperclip
